ProfileVC: setUpCacheSettings function was moved to presenter
